change settings naming to be orderbook yaml
