+++ crates/js_api/src/lib.rs
@@ -1,6 +1,6 @@
 #[cfg(target_family = "wasm")]
 pub mod gui;
 #[cfg(target_family = "wasm")]
-pub mod settings;
-#[cfg(target_family = "wasm")]
 pub mod subgraph;
+#[cfg(target_family = "wasm")]
+pub mod yaml;
